requirements.txt
requests~=2.26.0
PyYAML~=6.0
certifi~=2021.10.8
idna~=3.3
urllib3~=1.26.7